(#38) feat: 기존 이슈 cell 의 컴포넌트를 가지는 UIView 생성
What
1. cell 이 가지고 있던 title, label 등의 IBOutlet 을 모두 새로운 UIView 로 넣었습니다
2. 그리고 이 UIView 는 swipe 기능을 위해 사용되는 StackView 에 들어갈 예정입니다
